+++ src/color.rs
@@ -1,3 +1,8 @@
+use std;
+use xcb::Connection;
+use xcb::xproto;
+use failure::{Error, err_msg};
+
 #[derive(Clone, Copy, PartialEq)]
 pub struct RGB {
     pub r: u8,
@@ -31,5 +36,24 @@ fn test_compaction() {
     assert!(RGB::new(0x00, 0x00, 0x00).is_compactable());
     assert!(!RGB::new(0xf7, 0xf7, 0xf7).is_compactable());
     assert!(!RGB::new(0xff, 0xf7, 0xff).is_compactable());
+}
+
+pub fn window_color_at_point(conn: &Connection, window: xproto::Window, (x, y): (i16, i16))
+                         -> Result<RGB, Error> {
+    let reply = xproto::get_image(conn,
+                                  xproto::IMAGE_FORMAT_Z_PIXMAP as u8,
+                                  window,
+                                  x, y, 1, 1,
+                                  std::u32::MAX)
+        .get_reply()?;
+    if reply.depth() != 24 {
+        // TODO: Figure out what to do with these
+        return Err(err_msg("Unsupported color depth"));
+    }
+    let data = reply.data();
+    let r = data[2];
+    let g = data[1];
+    let b = data[0];
+    Ok(RGB::new(r, g, b))
 }
